This change brings the ARM64 and RiscV runtimes into sync with the amd64 and s390x runtimes which tag certain runtime functions as signal handlers (caml_call_realloc_stack, caml_call_gc, caml_c_call, caml_c_call_stack_args, caml_start_program and caml_runstack), for the purpose of displaying backtraces correctly in GDB on Linux. See [Retrofitting Effect Handlers onto OCaml](https://doi.org/10.1145/3453483.3454039) paper section 2.3 Stack Unwinding for further details.

With this change backtraces are displayed with `<signal handler called>` shown whenever GDB encounters these functions e.g.
```
Breakpoint 2, camlFib.entry () at fib.ml:15
15	let _ = main ()
(gdb) bt
#0  camlFib.entry () at fib.ml:15
#1  0x0000aaaaaaaef244 in caml_program ()
#2  <signal handler called>
#3  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:141
#4  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#5  0x0000aaaaaab4a760 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#6  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#7  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
```
and
```
#0  caml_alloc_stack (hval=281474573335040, hexn=281474573334568, heff=281474573334496) at runtime/fiber.c:226
#1  <signal handler called>
#2  0x0000aaaaaab2e7b4 in camlStdlib__Effect.match_with_496 () at effect.ml:83
#3  0x0000aaaaaaafb1e8 in camlEff.entry () at eff.ml:52
#4  0x0000aaaaaaaf7b34 in caml_program ()
#5  <signal handler called>
#6  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:141
#7  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#8  0x0000aaaaaab578b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#9  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#10 caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
```

NOTE that power.S is not consistent either but I am working on a fix for it.