@gasche 2024-06-17
There are some aspects of the code, only semi-related to your changes, that I find confusing.

`caml_call_gc` and `caml_c_call` are both functions and labels (placed in the function, after the prologue), other functions jump to the labels after their own prologue. But the handling of CFI_SIGNAL_FRAME is inconsistent:
- currently (in amd64.S) the directive is always placed in the function prologue, before the label, so other functions that jump to the label will not encounter it
- as a non-expert, my impression is that this choice is probably right for the `caml_c_call` label (which is only used to call `caml_raise_*` functions in failure cases of larger functions, and probably wrong for the `caml_call_gc` label (I suppose that `caml_alloc<n>` variants would also benefit from the subroutine)
- in your PR, you do the same as amd64 in arm64, but riscv places the directive after the `caml_call_gc` label (which is probably better?)

I am not sure if before-the-label or after-the-label is the right choice and would welcome confirmation (this might also deserve a comment in the source), but I would expect all backends to use the same strategy for each function.