@dustanddreams 2024-06-17
> I am not sure if before-the-label or after-the-label is the right choice and would welcome confirmation (this might also deserve a comment in the source), but I would expect all backends to use the same strategy for each function.

The cfi annotation should be put after the label.