@tmcgilchrist 2024-06-17
@dustanddreams are you satisfied with the changes as they stand? 

> I would expect all backends to use the same strategy for each function.

The backends should be consistent with each other except for on amd64 where there is an optimisation allowing tail-calling into caml_call_gc function (that @dustanddreams explained to me offline).