@kayceesrk 2024-06-21
The reason some of the frames were originally tagged as signal handler frames was that GDB cut the backtrace when the stacks did not grow towards 0 address, which happens when the stacks switch at external calls, callbacks and effect handlers. From the original PR message, it is not clear to me whether the stacks were indeed cut at those positions and this PR fixes them; the PR message only shows the effect of applying this patch. 

Also, this backtrace seems ill-formed

```
#3  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:141
#4  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#5  0x0000aaaaaab4a760 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#6  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#7  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
```

why does `caml_startup_exn` call `caml_main`? The backtrace at entry should look like this

```
#3  <signal handler called>
#4  0x000055fc66df1a26 in caml_startup_common (pooling=<optimised out>, argv=0x7ffc7d25fc68)
    at runtime/startup_nat.c:128
#5  caml_startup_common (argv=0x7ffc7d25fc68, pooling=<optimised out>) at runtime/startup_nat.c:87
#6  0x000055fc66df1a9f in caml_startup_exn (argv=<optimised out>) at runtime/startup_nat.c:135
#7  caml_startup (argv=<optimised out>) at runtime/startup_nat.c:140
#8  caml_main (argv=<optimised out>) at runtime/startup_nat.c:147
#9  0x000055fc66d9b872 in main (argc=<optimised out>, argv=<optimised out>) at runtime/main.c:37
```