@tmcgilchrist 2024-06-26
Thank you for the extra context @kayceesrk 

Using the meander example from Retrofitting effect handlers onto OCaml https://dl.acm.org/doi/10.1145/3453483.3454039
```shell
$ cat meander.ml 
external ocaml_to_c
         : unit -> int = "ocaml_to_c"
exception E1
exception E2
let c_to_ocaml () = raise E1
let _ = Callback.register
          "c_to_ocaml" c_to_ocaml
let omain () =
  try (* h1 *)
    try (* h2 *) ocaml_to_c ()
    with E2 -> 0
  with E1 -> 42
let _ = assert (omain () = 42)%   
```
```shell
 $ cat meander_c.c 
#include <caml/mlvalues.h>
#include <caml/callback.h>

value ocaml_to_c (value unit) {
    caml_callback(*caml_named_value
                  ("c_to_ocaml"), Val_unit);
    return Val_int(0);
}%                                                        
```
Compiled using `ocamlopt meander_c.c meander.ml -o meander.exe`

Before this change on Linux / ARM64 on Ubuntu 24.04:
``` shell
(gdb) run
Starting program: /home/tsmc/ocaml/meander.exe 

This GDB supports auto-downloading debuginfo from the following URLs:
  <https://debuginfod.ubuntu.com>
Enable debuginfod for this session? (y or [n]) y
Debuginfod has been enabled.
To make this setting permanent, add 'set debuginfod enabled on' to .gdbinit.
Downloading separate debug info for system-supplied DSO at 0xfffff7ffb000
[Thread debugging using libthread_db enabled]                                                                                                             
Using host libthread_db library "/lib/aarch64-linux-gnu/libthread_db.so.1".

Breakpoint 1, 0x0000aaaaaaabd168 in caml_program ()
(gdb) bt
#0  0x0000aaaaaaabd168 in caml_program ()
#1  <signal handler called>
#2  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:141
#3  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#4  0x0000aaaaaaae91b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#5  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#6  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#7  0x0000000000000001 in ?? ()
Backtrace stopped: previous frame identical to this frame (corrupt stack?)
(gdb) c
Continuing.

Breakpoint 2, ocaml_to_c (unit=1) at meander_c.c:5
5	    caml_callback(*caml_named_value
(gdb) bt
#0  ocaml_to_c (unit=1) at meander_c.c:5
#1  0x0000aaaaaaae9590 in caml_c_call ()
Backtrace stopped: previous frame inner to this frame (corrupt stack?)
(gdb) c
Continuing.

Breakpoint 3, camlMeander.c_to_ocaml_273 () at meander.ml:5
5	let c_to_ocaml () = raise E1
(gdb) bt
#0  camlMeander.c_to_ocaml_273 () at meander.ml:5
#1  <signal handler called>
#2  0x0000aaaaaab20d60 in camlMeander.7 ()
#3  0x0000aaaaaab20d60 in camlMeander.data_begin ()
Backtrace stopped: previous frame identical to this frame (corrupt stack?)
(gdb) info br
Num     Type           Disp Enb Address            What
1       breakpoint     keep y   0x0000aaaaaaabd168 <caml_program>
	breakpoint already hit 1 time
2       breakpoint     keep y   0x0000aaaaaaac3a80 in ocaml_to_c at meander_c.c:5
	breakpoint already hit 1 time
3       breakpoint     keep y   0x0000aaaaaaabda70 in camlMeander.c_to_ocaml_273 at meander.ml:5
	breakpoint already hit 1 time
(gdb) 
```
There are three issues with this gdb session:
1. the backtrace for `caml_program` shows an inaccurate backtrace that doesn't correspond to the program execution.
2. the jump from OCaml to C at breakpoint 2 `ocaml_to_c` is not displaying correctly 
3. the second jump from the callback to `c_to_ocaml_273` is not displaying correctly

After this change:

```shell
(gdb) run
Starting program: /home/tsmc/ocaml/meander.exe 

This GDB supports auto-downloading debuginfo from the following URLs:
  <https://debuginfod.ubuntu.com>
Enable debuginfod for this session? (y or [n]) y
Debuginfod has been enabled.
To make this setting permanent, add 'set debuginfod enabled on' to .gdbinit.
Downloading separate debug info for system-supplied DSO at 0xfffff7ffb000
[Thread debugging using libthread_db enabled]                                                                                                             
Using host libthread_db library "/lib/aarch64-linux-gnu/libthread_db.so.1".

Breakpoint 1, 0x0000aaaaaaabd168 in caml_program ()
(gdb) bt
#0  0x0000aaaaaaabd168 in caml_program ()
#1  <signal handler called>
#2  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:141
#3  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#4  0x0000aaaaaaae91b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#5  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#6  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#7  0x0000000000000001 in ?? ()
Backtrace stopped: previous frame identical to this frame (corrupt stack?)
(gdb) c
Continuing.

Breakpoint 2, ocaml_to_c (unit=1) at meander_c.c:5
5	    caml_callback(*caml_named_value
(gdb) bt
#0  ocaml_to_c (unit=1) at meander_c.c:5
#1  <signal handler called>
#2  0x0000aaaaaaabdab8 in camlMeander.omain_278 () at meander.ml:10
#3  0x0000aaaaaaabdc9c in camlMeander.entry () at meander.ml:13
#4  0x0000aaaaaaabd210 in caml_program ()
#5  <signal handler called>
#6  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:141
#7  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#8  0x0000aaaaaaae91b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#9  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#10 caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#11 0x0000000000000001 in ?? ()
Backtrace stopped: previous frame identical to this frame (corrupt stack?)
(gdb) c
Continuing.

Breakpoint 3, camlMeander.c_to_ocaml_273 () at meander.ml:5
5	let c_to_ocaml () = raise E1
(gdb) bt
#0  camlMeander.c_to_ocaml_273 () at meander.ml:5
#1  <signal handler called>
#2  0x0000aaaaaab20d60 in camlMeander.7 ()
#3  0x0000aaaaaab20d60 in camlMeander.data_begin ()
Backtrace stopped: previous frame identical to this frame (corrupt stack?)
```
The frames are not labelled as `<signal handler called>`  but there are inaccurate stack frames show for the breakpoints `caml_program` and `camlMeander.c_to_ocaml_273`. This matches the Linux/x86_64 GDB session below captured on Ubuntu 22.04 (minor issues with gdb not associating OCaml frames with source files).
```shell
(gdb) run
Starting program: /home/tsmc/ocaml/meander.exe 
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".

Breakpoint 3, 0x000055555556d980 in caml_program ()
(gdb) bt
#0  0x000055555556d980 in caml_program ()
#1  <signal handler called>
#2  0x0000555555599196 in caml_startup_common (pooling=<optimised out>, argv=0x7fffffffe518) at runtime/startup_nat.c:128
#3  caml_startup_common (argv=0x7fffffffe518, pooling=<optimised out>) at runtime/startup_nat.c:87
#4  0x000055555559920f in caml_startup_exn (argv=<optimised out>) at runtime/startup_nat.c:135
#5  caml_startup (argv=<optimised out>) at runtime/startup_nat.c:140
#6  caml_main (argv=<optimised out>) at runtime/startup_nat.c:147
#7  0x000055555556d882 in main (argc=<optimised out>, argv=<optimised out>) at runtime/main.c:37
(gdb) c
Continuing.

Breakpoint 1, 0x0000555555572e50 in ocaml_to_c ()
(gdb) bt
#0  0x0000555555572e50 in ocaml_to_c ()
#1  <signal handler called>
#2  0x000055555556e063 in camlMeander.omain_278 ()
#3  0x000055555556e1e2 in camlMeander.entry ()
#4  0x000055555556d9e3 in caml_program ()
#5  <signal handler called>
#6  0x0000555555599196 in caml_startup_common (pooling=<optimised out>, argv=0x7fffffffe518) at runtime/startup_nat.c:128
#7  caml_startup_common (argv=0x7fffffffe518, pooling=<optimised out>) at runtime/startup_nat.c:87
#8  0x000055555559920f in caml_startup_exn (argv=<optimised out>) at runtime/startup_nat.c:135
#9  caml_startup (argv=<optimised out>) at runtime/startup_nat.c:140
#10 caml_main (argv=<optimised out>) at runtime/startup_nat.c:147
#11 0x000055555556d882 in main (argc=<optimised out>, argv=<optimised out>) at runtime/main.c:37
(gdb) c
Continuing.

Breakpoint 2, 0x000055555556e010 in camlMeander.c_to_ocaml_273 ()
(gdb) bt
#0  0x000055555556e010 in camlMeander.c_to_ocaml_273 ()
#1  <signal handler called>
#2  0x0000555555574078 in caml_callback_exn (closure=<optimised out>, arg=<optimised out>) at runtime/callback.c:206
#3  0x000055555557467d in caml_callback (closure=<optimised out>, arg=<optimised out>) at runtime/callback.c:347
#4  0x0000555555572e71 in ocaml_to_c ()
#5  <signal handler called>
#6  0x000055555556e063 in camlMeander.omain_278 ()
#7  0x000055555556e1e2 in camlMeander.entry ()
#8  0x000055555556d9e3 in caml_program ()
#9  <signal handler called>
#10 0x0000555555599196 in caml_startup_common (pooling=<optimised out>, argv=0x7fffffffe518) at runtime/startup_nat.c:128
#11 caml_startup_common (argv=0x7fffffffe518, pooling=<optimised out>) at runtime/startup_nat.c:87
#12 0x000055555559920f in caml_startup_exn (argv=<optimised out>) at runtime/startup_nat.c:135
#13 caml_startup (argv=<optimised out>) at runtime/startup_nat.c:140
#14 caml_main (argv=<optimised out>) at runtime/startup_nat.c:147
#15 0x000055555556d882 in main (argc=<optimised out>, argv=<optimised out>) at runtime/main.c:37
(gdb) info br
Num     Type           Disp Enb Address            What
1       breakpoint     keep y   0x0000555555572e50 <ocaml_to_c>
	breakpoint already hit 1 time
2       breakpoint     keep y   0x000055555556e010 <camlMeander.c_to_ocaml_273>
	breakpoint already hit 1 time
3       breakpoint     keep y   0x000055555556d980 <caml_program>
	breakpoint already hit 1 time
```

There are still issues with inaccurate stack frames which requires a small fix to the CFI directives on ARM64 https://github.com/tmcgilchrist/ocaml/commit/761f333acff0ae291ab6e27803a746b6f9d8538f after that you get an equivalent gdb session to Linux / x86_64.
```shell
(gdb) run
Starting program: /home/tsmc/ocaml/meander.exe 

This GDB supports auto-downloading debuginfo from the following URLs:
  <https://debuginfod.ubuntu.com>
Enable debuginfod for this session? (y or [n]) y
Debuginfod has been enabled.
To make this setting permanent, add 'set debuginfod enabled on' to .gdbinit.
Downloading separate debug info for system-supplied DSO at 0xfffff7ffb000
[Thread debugging using libthread_db enabled]                                                                                                             
Using host libthread_db library "/lib/aarch64-linux-gnu/libthread_db.so.1".

Breakpoint 1, 0x0000aaaaaaabd168 in caml_program ()
(gdb) bt
#0  0x0000aaaaaaabd168 in caml_program ()
#1  <signal handler called>
#2  0x0000aaaaaaae9138 in caml_startup_common (pooling=0, argv=0xfffffffffc78) at runtime/startup_nat.c:128
#3  caml_startup_common (argv=0xfffffffffc78, pooling=0) at runtime/startup_nat.c:87
#4  0x0000aaaaaaae91b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#5  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#6  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#7  0x0000aaaaaaabcfd0 in main (argc=<optimized out>, argv=<optimized out>) at runtime/main.c:37
(gdb) c
Continuing.

Breakpoint 2, ocaml_to_c (unit=1) at meander_c.c:5
5	    caml_callback(*caml_named_value
(gdb) bt
#0  ocaml_to_c (unit=1) at meander_c.c:5
#1  <signal handler called>
#2  0x0000aaaaaaabdab8 in camlMeander.omain_278 () at meander.ml:10
#3  0x0000aaaaaaabdc9c in camlMeander.entry () at meander.ml:13
#4  0x0000aaaaaaabd210 in caml_program ()
#5  <signal handler called>
#6  0x0000aaaaaaae9138 in caml_startup_common (pooling=4, argv=0xaaaaaab24ac8) at runtime/startup_nat.c:128
#7  caml_startup_common (argv=0xaaaaaab24ac8, pooling=4) at runtime/startup_nat.c:87
#8  0x0000aaaaaaae91b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#9  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#10 caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#11 0x0000aaaaaaabcfd0 in main (argc=<optimized out>, argv=<optimized out>) at runtime/main.c:37
(gdb) c
Continuing.

Breakpoint 3, camlMeander.c_to_ocaml_273 () at meander.ml:5
5	let c_to_ocaml () = raise E1
(gdb) bt
#0  camlMeander.c_to_ocaml_273 () at meander.ml:5
#1  <signal handler called>
#2  0x0000aaaaaaac4d28 in caml_callback_exn (closure=<optimized out>, arg=<optimized out>, arg@entry=1) at runtime/callback.c:206
#3  0x0000aaaaaaac531c in caml_callback (closure=<optimized out>, arg=arg@entry=1) at runtime/callback.c:347
#4  0x0000aaaaaaac3aa0 in ocaml_to_c (unit=<optimized out>) at meander_c.c:5
#5  <signal handler called>
#6  0x0000aaaaaaabdab8 in camlMeander.omain_278 () at meander.ml:10
#7  0x0000aaaaaaabdc9c in camlMeander.entry () at meander.ml:13
#8  0x0000aaaaaaabd210 in caml_program ()
#9  <signal handler called>
#10 0x0000aaaaaaae9138 in caml_startup_common (pooling=4, argv=0xaaaaaab24ac8) at runtime/startup_nat.c:128
#11 caml_startup_common (argv=0xaaaaaab24ac8, pooling=4) at runtime/startup_nat.c:87
#12 0x0000aaaaaaae91b0 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#13 caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#14 caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#15 0x0000aaaaaaabcfd0 in main (argc=<optimized out>, argv=<optimized out>) at runtime/main.c:37
(gdb)
```