@tmcgilchrist 2024-06-26
For Linux / RiscV the situation is worse.

Before this change:
```shell
(gdb) run
Starting program: /home/tsmc/ocaml/meander.exe 
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/riscv64-linux-gnu/libthread_db.so.1".

Breakpoint 1, 0x0000002aaaac5350 in caml_program ()
(gdb) bt
#0  0x0000002aaaac5350 in caml_program ()
#1  0x0000002aaaae70a4 in caml_start_program ()
#2  0x0000000000000001 in ?? ()
(gdb) c
Continuing.

Breakpoint 3, 0x0000002aaaacb3f0 in ocaml_to_c ()
(gdb) bt
#0  0x0000002aaaacb3f0 in ocaml_to_c ()
#1  0x0000002aaaae6f98 in caml_c_call ()
#2  0x0000002aaab31878 in ?? ()
(gdb) c
Continuing.

Breakpoint 2, 0x0000002aaaac5c88 in camlMeander.c_to_ocaml_273 ()
(gdb) bt
#0  0x0000002aaaac5c88 in camlMeander.c_to_ocaml_273 ()
#1  0x0000002aaaae70a4 in caml_start_program ()
#2  0x0000002aaaac5ce0 in camlMeander.omain_278 ()
#3  0x0000000000000002 in ?? ()
(gdb) info br
Num     Type           Disp Enb Address            What
1       breakpoint     keep y   0x0000002aaaac5350 <caml_program+8>
	breakpoint already hit 1 time
2       breakpoint     keep y   0x0000002aaaac5c88 <camlMeander.c_to_ocaml_273+28>
	breakpoint already hit 1 time
3       breakpoint     keep y   0x0000002aaaacb3f0 <ocaml_to_c+12>
	breakpoint already hit 1 time
(gdb) 
```
After this change:
```
(gdb) run
Starting program: /home/tsmc/ocaml/meander.exe 
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/riscv64-linux-gnu/libthread_db.so.1".

Breakpoint 1, 0x0000002aaaac5350 in caml_program ()
(gdb) bt
#0  0x0000002aaaac5350 in caml_program ()
#1  <signal handler called>
#2  0x0000000000000001 in ?? ()
(gdb) br caml_startup_common
Breakpoint 5 at 0x2aaaae6990: caml_startup_common. (2 locations)
(gdb) run
The program being debugged has been started already.
Start it from the beginning? (y or n) y
Starting program: /home/tsmc/ocaml/meander.exe 
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/riscv64-linux-gnu/libthread_db.so.1".

Breakpoint 5.1, caml_startup_common (argv=0x3ffffff398, pooling=0) at runtime/startup_nat.c:88
88	{
(gdb) bt
#0  caml_startup_common (argv=0x3ffffff398, pooling=0) at runtime/startup_nat.c:88
#1  0x0000002aaaae6ab6 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#2  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#3  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#4  0x0000002aaaac528a in main (argc=<optimized out>, argv=<optimized out>) at runtime/main.c:37
(gdb) c
Continuing.

Breakpoint 5.2, caml_startup_common (pooling=<optimized out>, argv=0x3ffffff398) at runtime/startup_nat.c:100
100	  caml_init_codefrag();
(gdb) bt
#0  caml_startup_common (pooling=<optimized out>, argv=0x3ffffff398) at runtime/startup_nat.c:100
#1  caml_startup_common (argv=0x3ffffff398, pooling=<optimized out>) at runtime/startup_nat.c:87
#2  0x0000002aaaae6ab6 in caml_startup_exn (argv=<optimized out>) at runtime/startup_nat.c:135
#3  caml_startup (argv=<optimized out>) at runtime/startup_nat.c:140
#4  caml_main (argv=<optimized out>) at runtime/startup_nat.c:147
#5  0x0000002aaaac528a in main (argc=<optimized out>, argv=<optimized out>) at runtime/main.c:37
(gdb) c
Continuing.

Breakpoint 1, 0x0000002aaaac5350 in caml_program ()
(gdb) bt
#0  0x0000002aaaac5350 in caml_program ()
#1  <signal handler called>
#2  0x0000000000000001 in ?? ()
(gdb) c
Continuing.

Breakpoint 3, 0x0000002aaaacb3f0 in ocaml_to_c ()
(gdb) bt
#0  0x0000002aaaacb3f0 in ocaml_to_c ()
#1  <signal handler called>
#2  0x0000002aaab31878 in ?? ()
(gdb) info br
Num     Type           Disp Enb Address            What
1       breakpoint     keep y   0x0000002aaaac5350 <caml_program+8>
	breakpoint already hit 1 time
2       breakpoint     keep y   0x0000002aaaac5c88 <camlMeander.c_to_ocaml_273+28>
3       breakpoint     keep y   0x0000002aaaacb3f0 <ocaml_to_c+12>
	breakpoint already hit 1 time
4       breakpoint     keep y   0x0000002aaaac5350 <caml_program+8>
	breakpoint already hit 1 time
5       breakpoint     keep y   <MULTIPLE>         
	breakpoint already hit 2 times
5.1                         y   0x0000002aaaae6990 in caml_startup_common at runtime/startup_nat.c:88
5.2                         y   0x0000002aaaae69c6 in caml_startup_common at runtime/startup_nat.c:100
```
Small improvement as sections of the backtrace get labelled correctly with `<signal handler called>`  but the CFI information for this backend is clearly broken and we loose track of frames when initially transitioning from C runtime into OCaml. I haven't got a fix for this yet.